Make extractFields reusable
