+++ lib/app.js
@@ -2,6 +2,7 @@
 
 const request = require('./utils/request');
 const matchScriptData = require('./utils/matchScriptData');
+const generateExtractFieldsFn = require('./utils/generateExtractFieldsFn');
 const queryString = require('querystring');
 const cheerio = require('cheerio');
 const R = require('ramda');
@@ -32,7 +33,7 @@ function app (opts) {
     request(options, opts.throttle)
       .then(matchScriptData)
     // comment next line to get raw data
-      .then(extractFields)
+      .then(generateExtractFieldsFn(MAPPINGS))
       .then(R.assoc('appId', opts.appId))
       .then(R.assoc('url', reqUrl))
       .then(resolve)
@@ -137,23 +138,6 @@ const MAPPINGS = {
   }
 
 };
-
-/*
- * Map the MAPPINGS object, applying each field spec to the parsed data.
- * If the mapping value is an array, use it as the path to the extract the
- * field's value. If it's an object, extract the value in object.path and pass
- * it to the function in object.fun
- */
-function extractFields (parsedData) {
-  return R.map((spec) => {
-    if (R.is(Array, spec)) {
-      return R.path(spec, parsedData);
-    }
-    // assume spec object
-    const input = R.path(spec.path, parsedData);
-    return spec.fun(input);
-  }, MAPPINGS);
-}
 
 function descriptionText (description) {
   // preserve the line breaks when converting to text
